LayoutTests/platform/chromium/fast/events/intercept-postmessage-expected.txt

main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
intercepted postMessage
intercepted postMessage
RECV1
